requirements.txt
Django==1.11.20
python-decouple==3.1
mysqlclient==1.3.12
oauthlib==2.1.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
psycopg2
django-widget-tweaks==1.4.3
django-session-security
pandas
numpy
sklearn
elasticsearch
djangorestframework==3.6.4
keras==2.1.2
tensorflow